+++ PackedDataFormatInfo.cpp
@@ -7,13 +7,13 @@ std::vector<juce::uint8> PackedDataFormatInfo::applyMapping(std::vector<PackedDa
 	std::vector<uint8> aprBlock(sizeDestination, 0);
 	for (auto info : kToneFormatDefinition) {
 		// Read the value from the input block
-		uint8 value = (source[info.byteIndex] >> info.lsbIndex) & ((1 << info.bitCount) - 1);
+		uint8 value = (source[(size_t) info.byteIndex] >> info.lsbIndex) & ((1 << info.bitCount) - 1);
 		// Map it!
 		if (info.converter_) {
 			value = info.converter_(value);
 		}
 		// Poke it into the right place 
-		aprBlock[info.targetParameter] |= value << info.targetBitIndex;
+		aprBlock[(size_t) info.targetParameter] |= value << info.targetBitIndex;
 	}
 	return aprBlock;
 }
