+++ Sysex.cpp
@@ -91,7 +91,7 @@ std::vector<juce::MidiMessage> Sysex::vectorToMessages(std::vector<uint8> const 
 		int bytesUsed = 0;
 		//TODO - this crashes in case the data is not well formed sysex (example: Depeche Mode TI sound set, there is a message which ends on 0xff 0xff). 
 		messages.push_back(MidiMessage(&midiData[inPointer], (int)(midiData.size() - inPointer), bytesUsed, lastStatusByte, 0.0, false));
-		inPointer += bytesUsed;
+		inPointer += (size_t) bytesUsed;
 	}
 	return messages;
 }
@@ -105,7 +105,7 @@ std::vector<juce::MidiMessage> Sysex::memoryBlockToMessages(MemoryBlock const &m
 		int bytesUsed = 0;
 		//TODO - this crashes in case the data is not well formed sysex (example: Depeche Mode TI sound set, there is a message which ends on 0xff 0xff). 
 		messages.push_back(MidiMessage(&midiData[inPointer], (int)(midiData.getSize() - inPointer), bytesUsed, lastStatusByte, 0.0, false));
-		inPointer += bytesUsed;
+		inPointer += (size_t) bytesUsed;
 	}
 	return messages;
 }
@@ -121,7 +121,7 @@ void Sysex::saveSysex(std::string const &filename, std::vector<juce::MidiMessage
 	FileOutputStream outputStream(sysexFile);
 	if (sysexFile.existsAsFile()) {
 		for (auto message : messages) {
-			outputStream.write(message.getRawData(), message.getRawDataSize());
+			outputStream.write(message.getRawData(), (size_t) message.getRawDataSize());
 		}
 	}
 }
